feat(controller function): map to controller function now can fallback
It can fallback to path + method (path in lower and method in PascalCase) as function name if no
operationId is specified. This changes the default behavior! The current behavior can be achieved by
an opt-out.

BREAKING CHANGE: If no operationId is specified in an operation the mapped controller function name
falls back to path + method (path in lower and method in PascalCase). You can opt this out with
fallbackControllerFunctionToPath on AddFromSpecificationOpts.

Closes #18